# backend/swagger.yaml
openapi: 3.0.3
info:
  title: Olympia HR Platform API
  description: |
    API REST pour la plateforme intelligente de gestion des ressources humaines.
    Conforme à la réglementation tunisienne du travail et de la paie.
    
    ## Fonctionnalités
    - Gestion employés avec validation Joi
    - Présence biométrique anti-fraude
    - Heures supplémentaires (125%, 150%, 200%)
    - Paie conforme IRPP 2025 + CNSS
    - Congés workflow complet
    - Déclarations légales (CNSS, IR)
    - Analyse sentiment IA
    
  version: 1.0.0
  contact:
    name: Support Olympia HR
    email: support@olympia-hr.tn
servers:
  - url: http://localhost:5000/api
    description: Serveur de développement
  - url: https://api.olympia-hr.tn/api
    description: Serveur de production
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtenu via /auth/login
  schemas:
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "Erreur de validation"
        errors:
          type: array
          items:
            type: string
...
    Overtime:
      type: object
      required:
        - employee_id
        - date
        - hours
        - rate_type
        - reason
      properties:
        overtime_id:
          type: string
          example: "ot-456def"
        employee_id:
          type: string
          example: "emp-123abc"
        date:
          type: string
          format: date
          example: "2025-12-25"
        hours:
          type: number
          minimum: 0.5
          maximum: 12
          example: 4
        rate_type:
          type: string
          enum: ["125%", "150%", "200%"]
          example: "125%"
        overtime_category:
          type: string
          enum: [regular, night, sunday, holiday]
          example: "regular"
        reason:
          type: string
          maxLength: 500
          example: "Projet urgent à livrer"
        amount:
          type: number
          example: 85.23
        status:
          type: string
          enum: [pending, approved, rejected, cancelled]
          default: pending
...
paths:
...
  /overtime:
    get:
      tags:
        - Overtime
      summary: Liste des heures supplémentaires
      security:
        - bearerAuth: []
      parameters:
        - name: month
          in: query
          schema:
            type: string
            pattern: '^\d{4}-\d{2}$'
          example: "2025-12"
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, rejected]
        - name: employee_id
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Liste récupérée
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  count:
                    type: integer
                    example: 15
                  overtimes:
                    type: array
                    items:
                      $ref: '#/components/schemas/Overtime'
        '401':
          description: Non autorisé
    
    post:
      tags:
        - Overtime
      summary: Créer demande heures supplémentaires
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Overtime'
            example:
              employee_id: "emp-123abc"
              date: "2025-12-25"
              hours: 4
              rate_type: "125%"
              overtime_category: "regular"
              reason: "Projet urgent client"
      responses:
        '201':
          description: Demande créée
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Demande créée avec succès"
                  overtime:
                    $ref: '#/components/schemas/Overtime'
        '400':
          description: Validation échouée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /overtime/{id}/approve:
    put:
      tags:
        - Overtime
      summary: Approuver heures supplémentaires
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                manager_comments:
                  type: string
      responses:
        '200':
          description: Demande approuvée
        '403':
          description: Accès refusé (Manager/Admin seulement)
...